연결 정보 검색 로직 수정

+++ legacy/2021NIA/NiaIpPerfLinkageApplication/src/main/java/com/nia/data/linkage/ip/perf/mapper/linkage/LinkagePerfMapper.java
@@ -8,5 +8,5 @@ import java.util.ArrayList;
 @Mapper
 public interface LinkagePerfMapper {
     PerfVo selectMaxIntTimestamp();
-    ArrayList<PerfVo> selectPerfList(long intTimestamp);
+    ArrayList<PerfVo> selectPerfList(int intTimestamp);
 }
